feat: added method 'processImmediate' to synchronously process one event and get the result
There was no possibility to process synchronously one event in particular. the method 'canAccept' could be used to check if a transition would be accepted, but there was no guarantee that after the check the state machine was still in the same state. So this new method lets the user process the event synchronously and get the result of the processing; was the event accepted or not (no transition or rejected by guards).